+++ biobb_structure_checking/modelling/sequencedata.py
@@ -10,6 +10,7 @@ from Bio.PDB.Polypeptide import PPBuilder
 from Bio import SeqIO
 import sys
 import os
+import gzip
 # from typing import List, Dict
 from urllib.request import urlretrieve, urlcleanup
 
@@ -74,38 +75,46 @@ class SequenceData():
         """
         read_ok = True
         self.fasta = []
+        remove_fasta = False
         if fasta_sequence_path:
             if fasta_sequence_path.startswith('pdb:'):
                 fasta_sequence_path = fasta_sequence_path[4:]
                 try:
-                    tmp_file = f"/tmp/{fasta_sequence_path}.fasta"
+                    fasta_real_path = f"/tmp/{fasta_sequence_path}.fasta"
                     url = f"{FASTA_DOWNLOAD_PREFIX}/{fasta_sequence_path}"
                     print(f"Retrieving sequence from {url}")
                     urlcleanup()
-                    urlretrieve(url, tmp_file)
-                    for record in SeqIO.parse(tmp_file, 'fasta'):
-                        self.fasta.append(record)
-                    os.remove(tmp_file)
+                    urlretrieve(url, fasta_real_path)
+                    remove_fasta = True
                 except IOError:
                     sys.exit("Error retrieving FASTA")
+
             elif fasta_sequence_path.startswith('http'):
-                tmp_file = f'/tmp/{os.path.basename(fasta_sequence_path)}'
+                fasta_real_path = f'/tmp/{os.path.basename(fasta_sequence_path)}'
                 print(f"Downloading sequence from {fasta_sequence_path} ...")
                 try:
                     urlcleanup()
-                    urlretrieve(fasta_sequence_path, tmp_file)
-                    for record in SeqIO.parse(tmp_file, 'fasta'):
-                        self.fasta.append(record)
-                    os.remove(tmp_file)
+                    urlretrieve(fasta_sequence_path, fasta_real_path)
+                    remove_fasta = True
                 except IOError as e:
                     print(e)
-                    print(f"Error retrieving FASTA")
+                    print("Error retrieving FASTA")
             else:
-                try:
-                    for record in SeqIO.parse(fasta_sequence_path, 'fasta'):
-                        self.fasta.append(record)
-                except IOError:
-                    sys.exit("Error loading FASTA")
+                fasta_real_path = fasta_sequence_path
+
+            try:
+                if ".gz" in fasta_real_path:
+                    fasta_fh = gzip.open(fasta_real_path, 'rt')
+                else:
+                    fasta_fh = open(fasta_real_path, 'r')
+                for record in SeqIO.parse(fasta_fh, 'fasta'):
+                    self.fasta.append(record)
+            except IOError:
+                sys.exit(f"Error loading FASTA at {fasta_real_path}")
+
+            if remove_fasta:
+                os.remove(fasta_real_path)
+
         if not self.fasta:
             print(
                 f"WARNING: No valid FASTA formatted sequences found in"
@@ -176,8 +185,9 @@ class SequenceData():
                 if cif_warn:
                     print("Warning: sequence data unavailable on cif data")
                 return True
+
         for mod in strucm.st:
-            if mod not in self.data:
+            if mod.id not in self.data:
                 self.data[mod.id] = {}
             for index, chid in enumerate(chids):
                 for ch_id in chid.split(','):
@@ -205,7 +215,6 @@ class SequenceData():
                     self.data[mod.id][ch_id]['can'].features.append(
                         SeqFeature(FeatureLocation(1, len(seqs[index])))
                     )
-
                     # for chn in chids[i].split(','):
                     #     if chn in strucm.chain_ids:
                     #         self.data[ch_id]['chains'].append(chn)
@@ -219,7 +228,6 @@ class SequenceData():
                             if match[2] > IDENT_THRES * max_score:
                                 self.data[mod.id][ch_id]['chains'].append(match[1])
 
-            print(f"Canonical sequence for model {mod.id}:")
             self.has_canonical[mod.id] = {}
             for ch_id in strucm.chains_data.chain_ids[mod.id]:
                 self.has_canonical[mod.id][ch_id] = (ch_id in self.data[mod.id]) and\
@@ -288,7 +296,8 @@ class SequenceData():
                             frag,
                             strucm.chains_data.chain_ids[mod.id][chn.id]
                         )
-
+                    if isinstance(seq, str):
+                        seq = Seq(seq)
                     sqr = SeqRecord(
                         seq,
                         'pdbsq_' + frid,
@@ -412,6 +421,7 @@ class SequenceData():
                 )
                 self.data[mod.id][ch_id]['chains'].append(ch_id)
                 self.has_canonical[mod.id][ch_id] = True
+
 
     def get_canonical(self):
         """ SequenceData.get_canonical